troubleshoot

+++ utilities.hpp
@@ -8,5 +8,5 @@ vector diff(double x, double y, double z, int index);
 vector velocity(double x, double y, double z);
 double divvadvecv(double **** a, double x, double y, double z,int n, double h);
 vector*** divtenall(tensor*** a, const int n, double h);
-double partdif(long double**** a, int x, int y, int z, int n, double h, int d, int e);
+double partdif(double**** a, int x, int y, int z, int n, double h, int d, int e);
 #endif
