+++ vtk.hpp
@@ -6,8 +6,8 @@
 #include <iostream>
 
 // Original 3D version
-void writeVTKFile(int timestep, long double **** vp4d, int Nx, int Ny, int Nz, 
-                 long double dx, long double dy, long double dz);
+void writeVTKFile(int timestep, double **** vp4d, int Nx, int Ny, int Nz, 
+                 double dx, double dy, double dz);
 
 // New 2D version
 void writeVTKFile(const std::string& filename, const std::vector<std::vector<double>>& u,
